refactor: split long Vue files and fix frontend taste issues
- Extract MappingGroupFormDialog and MappingGroupDeleteDialog from ModelMappings.vue
- Extract useMetrics composable from Metrics.vue
- Replace v-html with lucide-vue-next icon components in Sidebar.vue
- Centralize API endpoint constants in client.ts
