+++ frontend/.gitignore
@@ -1,5 +1,5 @@
 # Logs
-/logs
+logs
 *.log
 npm-debug.log*
 yarn-debug.log*
