+++ frontend/src/components/layout/Sidebar.vue
@@ -1,4 +1,4 @@
-<!-- eslint-disable vue/multi-word-component-names vue/no-v-html -->
+<!-- eslint-disable vue/multi-word-component-names -->
 <template>
   <aside class="w-56 bg-slate-800 text-white flex-shrink-0 flex flex-col">
     <div class="p-4 border-b border-slate-700">
@@ -19,7 +19,7 @@
         class="flex items-center gap-3 px-3 py-2 rounded-lg text-sm transition-colors"
         :class="isActive(item.path) ? 'bg-slate-700 text-blue-400' : 'text-gray-300 hover:bg-slate-700'"
       >
-        <span v-html="item.icon"></span>
+        <component :is="item.icon" class="w-4 h-4" />
         {{ item.label }}
       </router-link>
     </nav>
@@ -29,9 +29,7 @@
         class="w-full justify-start text-gray-400 hover:text-white hover:bg-slate-700"
         @click="handleLogout"
       >
-        <svg class="w-4 h-4" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-          <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M17 16l4-4m0 0l-4-4m4 4H7m6 4v1a3 3 0 01-3 3H6a3 3 0 01-3-3V7a3 3 0 013-3h4a3 3 0 013 3v1"/>
-        </svg>
+        <LogOut class="w-4 h-4" />
         登出
       </Button>
     </div>
@@ -39,50 +37,39 @@
 </template>
 
 <script setup lang="ts">
+import { type Component, h } from 'vue'
 import { useRoute, useRouter } from 'vue-router'
+import {
+  LayoutDashboard,
+  Server,
+  BarChart3,
+  ArrowLeftRight,
+  KeyRound,
+  RefreshCcw,
+  FileText,
+  LogOut,
+} from 'lucide-vue-next'
 import { api } from '@/api/client'
 import { Button } from '@/components/ui/button'
 
+interface NavItem {
+  path: string
+  label: string
+  icon: Component
+}
+
+const navItems: NavItem[] = [
+  { path: '/', label: '仪表盘', icon: LayoutDashboard },
+  { path: '/providers', label: '供应商', icon: Server },
+  { path: '/metrics', label: '性能指标', icon: BarChart3 },
+  { path: '/mappings', label: '模型映射', icon: ArrowLeftRight },
+  { path: '/router-keys', label: 'API 密钥', icon: KeyRound },
+  { path: '/retry-rules', label: '重试规则', icon: RefreshCcw },
+  { path: '/logs', label: '请求日志', icon: FileText },
+]
+
 const route = useRoute()
 const router = useRouter()
-
-const navItems = [
-  {
-    path: '/',
-    label: '仪表盘',
-    icon: '<svg class="w-4 h-4" fill="none" stroke="currentColor" viewBox="0 0 24 24"><path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M4 6a2 2 0 012-2h2a2 2 0 012 2v2a2 2 0 01-2 2H6a2 2 0 01-2-2V6zM14 6a2 2 0 012-2h2a2 2 0 012 2v2a2 2 0 01-2 2h-2a2 2 0 01-2-2V6zM4 16a2 2 0 012-2h2a2 2 0 012 2v2a2 2 0 01-2 2H6a2 2 0 01-2-2v-2zM14 16a2 2 0 012-2h2a2 2 0 012 2v2a2 2 0 01-2 2h-2a2 2 0 01-2-2v-2z"/></svg>',
-  },
-  {
-    path: '/providers',
-    label: '供应商',
-    icon: '<svg class="w-4 h-4" fill="none" stroke="currentColor" viewBox="0 0 24 24"><path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M5 12h14M5 12a2 2 0 01-2-2V6a2 2 0 012-2h14a2 2 0 012 2v4a2 2 0 01-2 2M5 12a2 2 0 00-2 2v4a2 2 0 002 2h14a2 2 0 002-2v-4a2 2 0 00-2-2"/></svg>',
-  },
-  {
-    path: '/metrics',
-    label: '性能指标',
-    icon: '<svg class="w-4 h-4" fill="none" stroke="currentColor" viewBox="0 0 24 24"><path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M9 19v-6a2 2 0 00-2-2H5a2 2 0 00-2 2v6a2 2 0 002 2h2a2 2 0 002-2zm0 0V9a2 2 0 012-2h2a2 2 0 012 2v10m-6 0a2 2 0 002 2h2a2 2 0 002-2m0 0V5a2 2 0 012-2h2a2 2 0 012 2v14a2 2 0 01-2 2h-2a2 2 0 01-2-2z"/></svg>',
-  },
-  {
-    path: '/mappings',
-    label: '模型映射',
-    icon: '<svg class="w-4 h-4" fill="none" stroke="currentColor" viewBox="0 0 24 24"><path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M8 7h12m0 0l-4-4m4 4l-4 4m0 6H4m0 0l4 4m-4-4l4-4"/></svg>',
-  },
-  {
-    path: '/router-keys',
-    label: 'API 密钥',
-    icon: '<svg class="w-4 h-4" fill="none" stroke="currentColor" viewBox="0 0 24 24"><path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M15 7a2 2 0 012 2m4 0a6 6 0 01-7.743 5.743L11 17H9v2H7v2H4a1 1 0 01-1-1v-2.586a1 1 0 01.293-.707l5.964-5.964A6 6 0 1121 9z"/></svg>',
-  },
-  {
-    path: '/retry-rules',
-    label: '重试规则',
-    icon: '<svg class="w-4 h-4" fill="none" stroke="currentColor" viewBox="0 0 24 24"><path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M4 4v5h.582m15.356 2A8.001 8.001 0 004.582 9m0 0H9m11 11v-5h-.581m0 0a8.003 8.003 0 01-15.357-2m15.357 2H15"/></svg>',
-  },
-  {
-    path: '/logs',
-    label: '请求日志',
-    icon: '<svg class="w-4 h-4" fill="none" stroke="currentColor" viewBox="0 0 24 24"><path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M9 12h6m-6 4h6m2 5H7a2 2 0 01-2-2V5a2 2 0 012-2h5.586a1 1 0 01.707.293l5.414 5.414a1 1 0 01.293.707V19a2 2 0 01-2 2z"/></svg>',
-  },
-]
 
 function isActive(path: string): boolean {
   if (path === '/') return route.path === '/'
