+++ frontend/src/views/Logs.vue
@@ -112,7 +112,7 @@
               客户端原始请求
             </CollapsibleTrigger>
             <CollapsibleContent>
-              <LogRequestViewer :raw="detailData.client_request" :api-type="(detailData.api_type as 'openai' | 'anthropic') || 'anthropic'" />
+              <pre class="bg-gray-900 text-green-400 rounded-md p-3 text-xs overflow-auto max-h-[25vh] whitespace-pre-wrap break-all mt-1">{{ formatJson(detailData.client_request) }}</pre>
             </CollapsibleContent>
           </Collapsible>
           <Collapsible v-if="detailData.upstream_request" class="space-y-1">
@@ -121,7 +121,7 @@
               代理发送给 LLM API 的请求
             </CollapsibleTrigger>
             <CollapsibleContent>
-              <LogRequestViewer show-url :raw="detailData.upstream_request" :api-type="(detailData.api_type as 'openai' | 'anthropic') || 'anthropic'" />
+              <pre class="bg-gray-900 text-yellow-400 rounded-md p-3 text-xs overflow-auto max-h-[25vh] whitespace-pre-wrap break-all mt-1">{{ formatJson(detailData.upstream_request) }}</pre>
             </CollapsibleContent>
           </Collapsible>
           <Collapsible v-if="detailData.upstream_response" class="space-y-1">
@@ -130,7 +130,7 @@
               LLM API 返回的原始响应
             </CollapsibleTrigger>
             <CollapsibleContent>
-              <LogResponseViewer :raw="detailData.upstream_response" :api-type="(detailData.api_type as 'openai' | 'anthropic') || 'anthropic'" :is-stream="!!detailData.is_stream" />
+              <pre class="bg-gray-900 text-blue-400 rounded-md p-3 text-xs overflow-auto max-h-[25vh] whitespace-pre-wrap break-all mt-1">{{ formatJson(detailData.upstream_response) }}</pre>
             </CollapsibleContent>
           </Collapsible>
           <Collapsible v-if="detailData.client_response" class="space-y-1">
@@ -139,7 +139,7 @@
               代理返回给客户端的响应
             </CollapsibleTrigger>
             <CollapsibleContent>
-              <LogResponseViewer :raw="detailData.client_response" :api-type="(detailData.api_type as 'openai' | 'anthropic') || 'anthropic'" :is-stream="!!detailData.is_stream" />
+              <pre class="bg-gray-900 text-purple-400 rounded-md p-3 text-xs overflow-auto max-h-[25vh] whitespace-pre-wrap break-all mt-1">{{ formatJson(detailData.client_response) }}</pre>
             </CollapsibleContent>
           </Collapsible>
 
@@ -147,13 +147,11 @@
           <template v-if="!detailData.client_request">
             <div>
               <h3 class="text-sm font-medium text-gray-700 mb-1">Request Body</h3>
-              <LogRequestViewer v-if="detailData.request_body" :raw="detailData.request_body" :api-type="(detailData.api_type as 'openai' | 'anthropic') || 'anthropic'" />
-              <pre v-else class="bg-gray-900 text-green-400 rounded-md p-3 text-xs overflow-auto max-h-[25vh] whitespace-pre-wrap break-all">(无数据)</pre>
+              <pre class="bg-gray-900 text-green-400 rounded-md p-3 text-xs overflow-auto max-h-[25vh] whitespace-pre-wrap break-all">{{ formatJson(detailData.request_body) }}</pre>
             </div>
             <div>
               <h3 class="text-sm font-medium text-gray-700 mb-1">Response Body</h3>
-              <LogResponseViewer v-if="detailData.response_body" :raw="detailData.response_body" :api-type="(detailData.api_type as 'openai' | 'anthropic') || 'anthropic'" :is-stream="!!detailData.is_stream" />
-              <pre v-else class="bg-gray-900 text-blue-400 rounded-md p-3 text-xs overflow-auto max-h-[25vh] whitespace-pre-wrap break-all">(无数据)</pre>
+              <pre class="bg-gray-900 text-blue-400 rounded-md p-3 text-xs overflow-auto max-h-[25vh] whitespace-pre-wrap break-all">{{ formatJson(detailData.response_body) }}</pre>
             </div>
           </template>
           <div v-if="detailData.error_message">
@@ -197,8 +195,6 @@ import { Select, SelectTrigger, SelectValue, SelectContent, SelectItem } from '@
 import { Table, TableHeader, TableBody, TableRow, TableHead, TableCell } from '@/components/ui/table'
 import { Dialog, DialogContent, DialogScrollContent, DialogHeader, DialogTitle, DialogFooter } from '@/components/ui/dialog'
 import { Collapsible, CollapsibleContent, CollapsibleTrigger } from '@/components/ui/collapsible'
-import LogRequestViewer from '@/components/logs/LogRequestViewer.vue'
-import LogResponseViewer from '@/components/logs/LogResponseViewer.vue'
 
 interface LogEntry {
   id: string
@@ -229,6 +225,15 @@ const cleanupDays = ref(30) // eslint-disable-line no-magic-numbers
 const showDetail = ref(false)
 const detailLoading = ref(false)
 const detailData = ref<LogEntry | null>(null)
+
+function formatJson(raw: string | null): string {
+  if (!raw) return '(无数据)'
+  try {
+    return JSON.stringify(JSON.parse(raw), null, 2) // eslint-disable-line no-magic-numbers
+  } catch {
+    return raw
+  }
+}
 
 async function openDetail(id: string) {
   showDetail.value = true
